requirements.txt
aerich==0.9.2
aiocache==0.12.3
aiocron==2.1
aiofiles==25.1.0
aiosmtplib==4.0.2
asyncpg==0.31.0
beautifulsoup4==4.14.3
boto3==1.42.5
botocore==1.42.5
celery==5.6.0
databases==0.9.0
fastapi==0.124.0
fastapi-cache2==0.2.2
fastapi-limiter==0.1.6
fastapi-login==1.10.2
fastapi-mail==1.5.8
fastapi-pagination==0.15.2
ffmpeg-python==0.2.0
htmlmin==0.1.12
httpx==0.28.1
hypercorn==0.18.0
imagekitio==4.2.0
itsdangerous==2.2.0
numpy==2.3.5
pandas==2.3.3
passlib==1.7.4
pydantic==2.12.5
pydantic-settings==2.12.0
pyfcm==1.5.4
PyJWT==2.10.1
python-dotenv==1.2.1
python-slugify==8.0.4
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
pyvimeo==1.1.2
razorpay==2.0.0
redis==7.1.0
requests==2.32.5
slowapi==0.1.9
SQLAlchemy==2.0.44
SQLAlchemy-Utils==0.42.0
starlette==0.50.0
starlette-context==0.4.0
tortoise-orm==0.25.1
uvicorn==0.38.0
